pyproject.toml
[tool.poetry]
name = "unlearning-with-trak"
version = "0.1.0"
description = ""
authors = [""]

readme = "README.md"


packages = [
	{include = "unlearning"},
	{include = "plotting"},
	#{include = "trak"}
]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.4"
torch = "^2.2.1"
ipykernel = "^6.29.2"
matplotlib = "^3.8.3"
tqdm = "^4.66.2"
torchvision = "^0.17.1"
scikit-learn = "^1.4.1.post1"
pandas = "^2.2.1"
traker = "^0.3.2"
pickleshare = "^0.7.5"
seaborn = "^0.13.2"
jupyter-black = "^0.3.4"
opacus = "^1.4.1"
submitit = "^1.5.1"
transformers = "^4.40.2"
tiktoken = "^0.6.0"


# terminaltables
# fastargs

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
yapf = "^0.40.2"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
